// types/settings.ts
export interface Settings {
  colors: {
    primaryColor: string;
    primaryForeground: string;
    ring: string;
    background?: string;
    foreground?: string;
    secondaryColor: string;
    card: string;
    accent: string;
  };
  companyName: string;
  logoUrl: string;
  currency: string;
}